requirements.txt
Django==5.1.2
pymysql==1.1.0
python-dotenv==1.0.1
django-filter==24.3
whitenoise==6.7.0
gunicorn==22.0.0
